Allow passing session password in query string
Use a query string parameter called `pass` with a value of the session's password.
If trying to join a session, the password will be checked automatically and if it is correct, the user will be let in automatically.
If creating a session, the specified password will be automatically set as the new session's password.