Include icons in files as Base64 strings
The number of HTTP requests was getting pretty high, and the files are pretty small, so just include them in the source code.
`/public/img/*` has been moved to `/src/img/*`, except for `pen-icon-og.png`, which is used for the OpenGraph image and shouldn't be converted to a Base64 string.
It has, however, been moved from `/public/img/` to `/public/` because a folder for images isn't necessary when there's only 1.