Adjust icons for dark theme
The fully-black up and down arrows used in the action history table were difficult to see in the dark theme, so they have been outlined with white. They can now be seen easily in both light and dark themes.
Added a recoloured version of the external link icon to match the colour of links with the dark theme.