Skip disabled actions when undoing/redoing
If the target action in history when undoing or redoing is disabled, skip it, and try the next. Also make sure never to skip out of bounds of the actions list.
Enable and disable the undo/redo buttons based on whether there are any enabled actions in the past or future rather than just if the current position is at either one of the ends.