Trim fixarray markers from extension types
They're always going to be fixarrays, so the first byte can be removed
from the ext's data and added back when decoding.

The `PastAction` object can vary in length (2 or 3 items, `data` is not
always necessary, e.g. `CLEAR_CANVAS`), but it is smaller to have no
`data` property become a third item of `null` rather than to always save
the number of properties in each action.