Remove now unnecessary Canvas.update() second draw
The canvas used to be updated twice when "saving": once for drawing
everything but selections (since they're not part of the actual canvas
image) to copy over and save to the session canvas, and once for
everything including selections for display. Since selections are now
drawn onto the display canvas rather than the mixing canvas, this is no
longer necessary.