Add eraser tool
Although it just ends up being a Stroke, it might be useful to have
saved settings for erasing things (e.g. different size). It might also
be good to have for users who don't know about the "Erase" Mode on the
pen tool, or to make switching to erasing easier.