Draw selection box lines individually
Using ctx.strokeRect() with a dashed line means that the dashes of some
sides of the rectangle are affected by the others; the offset of the
dashes is affected by the lengths of the sides.

To avoid this, draw each side of the selection box individually,
resetting the dashes for each side.